Move tests to a better package
